feat(bot): Add guildId to ThreadMember
This uses an additional `extra` parameter in the transformer,
this is inconsistent with other transformers, as some use a custom object
as the payload and in there we have the discord payload + the additional
data we may need, like guildId in channel to make an example, while others
add additional parameters like the member transformer that has 2 additional
parameters for the guild id and the user id.

We should unity the approch we have to pass additional parameters to
transformers and customizers, so for this PR an `extra` parameter is used
and this edit will be added to the other transformers in another PR after
https://github.com/discordeno/discordeno/pull/4302 is merged as that pr
helps with allowing proper typechecking of the transformers and customizers,
along side not requiring the InternalBot "hack" to make everything work.