docs(bigbot): Send snake_case packets to bot (#4430)
Currently the big bot gateway worker was sending camelCase packets to the bot which caused the transformers in `@discordeno/bot` to not find most of the properties they need in the payload.

This changes the docs to show to override the forwardToBot method to send the packets as-is to the bot.

Co-authored-by: Awesome Stickz <awesome@stickz.dev>